adaptive_rl_blockchain/requirements.txt
simpy>=4.0
numpy>=1.23,<1.27
torch>=2.0,<2.2
pandas>=1.5,<2.2
matplotlib>=3.5,<4.0
hydra-core>=1.3,<2.0
tqdm>=4.60
PyYAML>=6.0
